+++ src/test/scala/edu/ie3/simona/service/weather/WeatherSourceSpec.scala
@@ -26,6 +26,7 @@ import edu.ie3.simona.test.common.UnitSpec
 import edu.ie3.util.geo.{CoordinateDistance, GeoUtils}
 import edu.ie3.util.quantities.QuantityUtil
 import org.locationtech.jts.geom.{Envelope, Point}
+import org.scalatest.prop.TableDrivenPropertyChecks
 import tech.units.indriya.ComparableQuantity
 import tech.units.indriya.quantity.Quantities
 import tech.units.indriya.unit.Units
@@ -37,7 +38,7 @@ import scala.jdk.CollectionConverters._
 import scala.jdk.OptionConverters._
 import scala.util.{Failure, Success, Try}
 
-class WeatherSourceSpec extends UnitSpec {
+class WeatherSourceSpec extends UnitSpec with TableDrivenPropertyChecks {
   private val coordinate0 = GeoUtils.buildPoint(51.47, 7.41)
 
   "A weather source" should {